+++ src/UseCase/IssuedInvoices/NewInvoice/Response/DeliveryAddress.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DobryProgramator\iDoklad\UseCase\IssuedInvoice\NewInvoice\Response;
+namespace DobryProgramator\iDoklad\UseCase\IssuedInvoices\NewInvoice\Response;
 
 use DobryProgramator\iDoklad\Enum\Country;
 use JMS\Serializer\Annotation as Serializer;
